Allow extendclass generator to detect factories.
